examples/python/RAG_QA_chatbot/pyproject.toml
[project]
name = "rag-qa-chatbot"
version = "0.1.0"
description = "RAG chatbot for documentation Q&A"
requires-python = ">=3.11"
dependencies = [
    # Ingestion
    "qdrant-client>=1.7.0",
    "litellm>=1.0.0",
    "python-frontmatter>=1.0.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    # Backend
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "sse-starlette>=2.0.0",
    # Observability
    "agenta>=0.80.0",
    "openinference-instrumentation-litellm>=0.1.28",
    "deepeval>=3.8.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ingest", "backend"]
